requirements.txt
google-genai>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
feedparser>=6.0.0
wikipedia-api>=0.7.0
pyyaml>=6.0.0
requests>=2.28.0
pydub>=0.25.0
